tcg: Introduce guard_{i32, i64} opcode
Opcode aims at checking certain assumption made at compile time.
It simply does a comparison on its 2 input argument and bailout
if the comparison fails. Note that the slow path never returns,
so the opcode itself is not tagged with TCG_OPF_CALL_CLOBBER.
Not used for the time being.

Signed-off-by: Liren Wei <lrwei@bupt.edu.cn>