tcg: Fix bug in tcg_gen_extract_tag
With speculation enabled, guard opcode now checks 3 things
at the same time:
1. Whether the access is aligned?
2. Whether the access crosses page boundary?
3. Whether the speculation is correct?
The way the old mechanism checks 2. is to pad the address
to that of the last byte of the access, which may cause a
false speculation to pass the guard, and in turn generate
a bad memory access.

Signed-off-by: Liren Wei <lrwei@bupt.edu.cn>