tcg: Adjust tlb_check to clobber less registers
Slow path of tlb_check SHOULD (through it hasn't been yet) be called
infrequently, adjust calling convention used in tlb_check to clobber
less registers on fast path. For x86_64 Linux host, mark r8 - r11 as
slow-path-saved-registers, Windows hosts are not supported yet.

Be aware that the actual tweaking of calling convention is quite hacky
in the sense that it's unsupported by Clang, and is documented as not
supported by GCC, though it seems to work under GCC 10.2.0.

Signed-off-by: Liren Wei <lrwei@bupt.edu.cn>