tcg: Introduce CLOBBER_ARG field to TCGLifeData
Like DEAD_ARG that was used to indicate the death of an output
temporary so that the register it uses can be released earlier,
CLOBBER_ARG marks those temporaries which might be clobbered by
CALL CLOBBER op and free the corresponding register at the last
use before the actual clobber instruction.
The new layout of TCGLifeData assumes 64-bit host and at most 7
argument for an op.

Signed-off-by: Liren Wei <lrwei@bupt.edu.cn>