tcg: Introduce CLOBBER_ARG field to TCGLifeData
Like DEAD_ARG that was used to indicate the death of an output
temporary so that the register it uses can be released earlier,
CLOBBER_ARG marks those temporaries which might be clobbered by
CALL_CLOBBER ops and free the corresponding register at the last
use before the actual clobber instruction.
The new layout of TCGLifeData assumes 64-bit host and at most 7
argument for an op.

This is introduced to avoid the following situation, which is not
unusual as most constant-additions are transformed to be directly
done on the base value.

```
    Assign %rdi to temporary x0

    ...
    (%rdi hold by temporary x0, alive, but not used before clobbered)
    ...

    Clobber register %rdi -- the actual release of %rdi here

    Use temporary x0
```

Signed-off-by: Liren Wei <lrwei@bupt.edu.cn>