tcg: Rename guard as guardm
After commit "tcg: Hardcode memory operand for tlb_check and guard",
guard_{i32, i64} can no longer function as a general-purpose checker.
And because the only way of generating the opcode is from tlb_check,
which, does not have both i32 and i64 versions, so rename it to be
"guardm" to clarify these things up.

Signed-off-by: Liren Wei <lrwei@bupt.edu.cn>