tcg: Break (and drop) support for TCG_TYPE_I64 on 32-bit hosts
Get tired of it, so get rid of it. Also remove TCGTemp::base_type field,
this is safe because there is really no user left after dropping support
for oversized guests. TCG vector code uses it, kind of for compatibility
with possible future supporting of "oversized vector temporary", but no
such things is available now.
Removing of .BASE_TYPE field is done by making it an alias of the .TYPE
field, so other code still get the illusion of dealing with temporaries
directly implemented by host hardware, i.e. with the 2 fields equal.

Signed-off-by: Liren Wei <lrwei@bupt.edu.cn>