tcg: Introduce and implement ITLB_CHECK
Opcode following similiar procedure of that of TLB_CHECK, in the sense
that they both RETURN on situations handleable within their slow path.
This is very different a design decision from that of HQEMU, where iTLB
miss simply causes bailout, and iTLB entries are updated after every TB
translation.

Within the slowpath of ITLB_CHECK, detected change of memory mapping
will trigger invalidation of the CURRENT TB, i.e. the TB to which the
ITLB_CHECK belongs to. This prevents branches towards WRONG / INVALID
TBs from continuing, in which, WRONG means (in)direct branches across
page, with the target virtual page now being mapped to another piece
of physical memory; and INVALID means TBs that are cross-page, being
invalidated also due to partially updated memory mapping.

Because of the possibility of "jump into slow path and return", the
opcode clobbers registers listed in tcg_target_tlb_check_clobber_regs,
as with TLB_CHECK.

Not emitted by the frontend for the time being.

Signed-off-by: Liren Wei <lrwei@bupt.edu.cn>