accel/tcg: Introduce cpu_really_has_work()
TCG trace recording mode will be interrupted if the execution is to be
redirected to target-specific interrupt handling routines. However, as
interrupt uses the same mechanism (icount_decr) to force TB exit early
for the pending interrupt to be able to be taken in time, we have to
distinguish those cases.

The original cpu_has_work() does not take care of masked interrupt for
certain target architecture, and may cause trace mode to be constantly
stopped. Check masked interrupt explicitly for those targets to avoid
this.

Signed-off-by: Liren Wei <lrwei@bupt.edu.cn>